Initial

+++ app/Http/Controllers/UserManagementController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Course;
+use App\Models\Logs;
 use App\Models\User;
 use App\Models\Year;
 use Auth;
@@ -43,8 +44,8 @@ class UserManagementController extends Controller
         if($yearID == 0 || $yearID == null){
             $courseID = null;
         }
-  
-
+        
+        $authUser = Auth::user();
         $user = User::create([
             'name' => $validated['name'],
             'email' => $validated['email'],
@@ -55,6 +56,13 @@ class UserManagementController extends Controller
             'yearID' => $yearID, 
             'is_verified' => 1,
             'avatar' => $avatarPath,
+        ]);
+
+        
+        Logs::create([
+            'userID' => $authUser->userID,
+            'action_type' => "Added a new user: {$validated['name']}.",
+            'timestamp' => now(),
         ]);
 
 
@@ -130,6 +138,15 @@ class UserManagementController extends Controller
     
         // Save the updated user
         $user->save();
+
+        $authUser = Auth::user();
+
+        Logs::create([
+            'userID' => $authUser->userID,
+            'action_type' => "Updated a user: {$validated['name']}.",
+            'timestamp' => now(),
+        ]);
+
     
         // Redirect with success message
         return redirect()->route('admin.user_management')->with('success', 'User updated successfully.');
@@ -144,11 +161,18 @@ class UserManagementController extends Controller
             return redirect()->route('admin.user_management')->with('error', 'Admin cannot be deleted.');
         }
 
-        // if ($userToDelete->id === $currentUser->id) {
-        //     return redirect()->route('admin.user_management')->with('error', 'You cannot delete your own account.');
-        // }
+        if ($userToDelete->id === $currentUser->id) {
+            return redirect()->route('admin.user_management')->with('error', 'You cannot delete your own account.');
+        }
 
         $userToDelete->delete();
+
+        Logs::create([
+            'userID' => $currentUser->userID,
+            'action_type' => "Deleted a user: {$userToDelete->name}.",
+            'timestamp' => now(),
+        ]);
+
 
 
         return redirect()->route('admin.user_management')->with('success', 'User has been deleted successfully.');
@@ -194,6 +218,12 @@ class UserManagementController extends Controller
         }
 
         $user->save();
+
+        Logs::create([
+            'userID' => $user->userID,
+            'action_type' => "Updated own profile of user: {$user->name}.",
+            'timestamp' => now(),
+        ]);
 
 
         return redirect()->route('profile.edit_profile')->with('success', 'User profile updated successfully.');
